flytaskpymdb/requirements.txt
asgiref==3.8.1
bcrypt==4.1.3
sqlparse==0.2.4
cryptography==42.0.7
django
django-cors-headers==4.3.1
djangorestframework==3.12.1
PyJWT==2.8.0
djongo
pymongo==3.12.0
dnspython==1.16.0
pytz
